requirements.txt
streamlit
pandas
fpdf
gspread
google-auth
google-api-python-client